@@ -5,7 +5,7 @@ import (
 	"log"
 	"strings"
 
-	. "github.com/outscale-dev/frieza/internal/common"
+	. "github.com/outscale/frieza/internal/common"
 	"github.com/teris-io/cli"
 )
 
